Switch to cvmUserId and cvmPassword instead of user-id and password
user-id is a variable in a weird format which limits usability in
certain situations due to syntax limitations.

Additionally, user-id and password are generic, and may clash with other
projects when a developer works with code from different organziations.

Prefix the variable that is in use with cvm to make it specific to
Central Valley Modeling, enabling a developer to work with multiple
different proejcts in parallel.

Additionally, allow compatibility with the legacy user-id and password
variables, but provide a warning about it to give other developers time
to update.